vps-services/pr-pipeline/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.40.0
tavily-python>=0.5.0
google-api-python-client>=2.0.0
requests>=2.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
